refactor scss and add icon styling to menu items
