+++ src/common/components/cards/instance-details-footer.scss
@@ -31,15 +31,4 @@
             margin: 14px 0px;
         }
     }
-
-    .kebab-menu-button {
-        background: transparent;
-        top: calc(50% - 4px / 2);
-
-        i {
-            font-size: 20px;
-            border: 1.25px;
-            color: $primary-text;
-        }
-    }
 }
